+++ crates/prover-clients/atlantic-service/tests/lib.rs
@@ -1,9 +1,10 @@
-use atlantic_service::AtlanticProverService;
+use atlantic_service::{AtlanticProverService, AtlanticValidatedArgs};
 use cairo_vm::types::layout_name::LayoutName;
 use cairo_vm::vm::runners::cairo_pie::CairoPie;
 use httpmock::MockServer;
 use prover_client_interface::{ProverClient, Task};
-use utils::settings::env::EnvSettingsProvider;
+use url::Url;
+use utils::env_utils::get_env_var_or_panic;
 
 use crate::constants::CAIRO_PIE_PATH;
 
@@ -14,7 +15,17 @@ async fn atlantic_client_submit_task_calls_correct_endpoint() {
     let _ = env_logger::try_init();
     color_eyre::install().expect("Unable to install color_eyre");
     dotenvy::from_filename("../.env.test").expect("Failed to load the .env file");
-
+    let atlantic_params = AtlanticValidatedArgs {
+        atlantic_api_key: get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_API_KEY"),
+        atlantic_service_url: Url::parse(&get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_SERVICE_URL")).unwrap(),
+        atlantic_rpc_node_url: Url::parse(&get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_RPC_NODE_URL")).unwrap(),
+        atlantic_mock_fact_hash: get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_MOCK_FACT_HASH"),
+        atlantic_prover_type: get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_PROVER_TYPE"),
+        atlantic_settlement_layer: get_env_var_or_panic("MADARA_ORCHESTRATOR_ATLANTIC_SETTLEMENT_LAYER"),
+        atlantic_verifier_contract_address: get_env_var_or_panic(
+            "MADARA_ORCHESTRATOR_ATLANTIC_VERIFIER_CONTRACT_ADDRESS",
+        ),
+    };
     // Start a mock server
     let mock_server = MockServer::start();
 
@@ -27,8 +38,7 @@ async fn atlantic_client_submit_task_calls_correct_endpoint() {
     });
 
     // Configure the service to use mock server
-    let settings = EnvSettingsProvider {};
-    let atlantic_service = AtlanticProverService::with_test_settings(&settings, mock_server.port());
+    let atlantic_service = AtlanticProverService::with_test_params(mock_server.port(), &atlantic_params);
 
     let cairo_pie_path = env!("CARGO_MANIFEST_DIR").to_string() + CAIRO_PIE_PATH;
     let cairo_pie = CairoPie::read_zip_file(cairo_pie_path.as_ref()).expect("failed to read cairo pie zip");